Enable `npm install` of forks of fedwiki/wiki-client
We create a workflow to automatically publish a node package to
github's package registry when a programmer pushes code to a branch
named with a prefix of `preview/`.

Developing this workflow was arduous. We're squashing 34 commits. So
keeping some notes here of things we learned along the way.

GitHub validates the yaml in actions, but at least some of the
validations do not provide useful error messages. Editing a yaml
workflow in the github web UI provided better feedback on the
validations.

Therefore, if changes to this workflow are not doing what you expect,
consider editing the file within the github web UI to see if there are
better error messages about the structure of the yaml itself.

with: and run: are mutually exclusive. It seems `with:` is intended
to supply parameters for packaged actions.

GitHub's registry is particular about how and what gets published. It
took a long time to understand these details. This conversation in
github.community was helpful. We paraphrase some specifics below in
case the link breaks in the future.
https://github.community/t/installing-npm-packages-from-the-github-package-registry/16150/6

- npm publish needs the repository and authentication token correctly
  configured in ~/.npmrc on the worker VM.

- actions/setup-node@v1 is the tool that updates the .npmrc file

Some combination of GitHub and npm publish require specific structure
and values in package.json:

- name must be scoped with the github organization of the repo
- repository.url must point to the fork'd repo, not the upstream repo
- packageConfig prevents `npm package` from publishing to npm's registry

We encapsulated the above three details in scripts/scopify-package
It's a relatively simple jq script to transform package.json, but it's
expanded into a longer bash script to clarify what each clause does
and why, and to provide usage for the script.

I also found out how to invoke `npm version` without git commits.